Bluetooth: ath3k: workaround the compatibility issue with xHCI controller

+++ crates/core_simd/tests/u8_ops.rs
@@ -1,5 +1,6 @@
+
 #![feature(portable_simd)]
 
 #[macro_use]
 mod ops_macros;
-impl_unsigned_tests! { u64 }
+impl_unsigned_tests! { u8 }
